Rename realtime service
