+++ .env.development
@@ -2,6 +2,8 @@ CLIENT_NAME=entur-tavla-staging
 JOURNEYPLANNER_HOST=https://api.staging.entur.io/journey-planner/v2
 GEOCODER_HOST=https://api.staging.entur.io/geocoder/v1
 MOBILITY_HOST = https://api.staging.entur.io/mobility/v2
+VEHICLES_REALTIME_HOST=https://api.staging.entur.io/realtime/v1/vehicles/graphql
+VEHICLES_REALTIME_SUBSCRIPTIONS_HOST=wss://api.staging.entur.io/realtime/v1/vehicles/subscriptions
 FIREBASE_CONFIG='{"apiKey":"AIzaSyDXpgZUj0IXBP3ECSlEk6IJwo_6VkYC_VE","authDomain":"entur-tavla-staging.firebaseapp.com","databaseURL":"https://entur-tavla-staging.firebaseio.com","projectId":"entur-tavla-staging","storageBucket":"entur-tavla-staging.appspot.com","messagingSenderId":"934908160432","appId":"1:934908160432:web:68c802afea624557ca4e75","measurementId":"G-1Z1FR2YGDZ"}'
 MAPBOX_TOKEN=pk.eyJ1IjoiZW50dXIiLCJhIjoiY2tlbGQ5MWpyMGloNjJ5cHd2bWY1dnRobyJ9.aWjUEMwRIIVIp1bF2Liyzw
 MAPBOX_STYLE_MAPVIEW=mapbox://styles/entur/ckfi7v87704jn19o71b6z02bp
