+++ .env.prod
@@ -2,6 +2,8 @@ CLIENT_NAME=entur-tavla
 JOURNEYPLANNER_HOST=https://api.entur.io/journey-planner/v2
 GEOCODER_HOST=https://api.entur.io/geocoder/v1
 MOBILITY_HOST = https://api.entur.io/mobility/v2
+VEHICLES_REALTIME_HOST=https://api.entur.io/realtime/v1/vehicles/graphql
+VEHICLES_REALTIME_SUBSCRIPTIONS_HOST=wss://api.entur.io/realtime/v1/vehicles/subscriptions
 FIREBASE_CONFIG='{"apiKey":"AIzaSyDBzat6zxRmQQQQsEpI3YiZ9uX4V9d-Bgg","authDomain":"entur-tavla-prod.firebaseapp.com","databaseURL":"https://entur-tavla-prod.firebaseio.com","projectId":"entur-tavla-prod","storageBucket":"entur-tavla-prod.appspot.com","messagingSenderId":"1025677335075","appId":"1:1025677335075:web:7e5e1f495564c1a0dcef96","measurementId":"G-GJLJPD7LYB"}'
 MAPBOX_TOKEN=pk.eyJ1IjoiZW50dXIiLCJhIjoiY2tlbGQ1OHQzMG10dTM1azdvOXJvcXExeCJ9.s65sEHnbUfmuqRsDwjByxw
 MAPBOX_STYLE_MAPVIEW=mapbox://styles/entur/ckfi7v87704jn19o71b6z02bp
