+++ src/services/realtimeVehicles/graphql.ts
