+++ src/services/realtimeVehicles/model/operator.ts
