+++ src/services/realtimeVehicles/model/statistics.ts
