+++ src/services/realtimeVehicles/model/vehicle.ts
