+++ src/services/realtimeVehicles/model/vehicleMapPoint.ts
