+++ src/services/realtimeVehicles/model/vehicleMode.ts
